feat: Implement graph builder

+++ src/Rulesage.Shared/Repositories/Implementations/NodeRepository.cs
@@ -53,7 +53,7 @@ public class NodeRepository(NpgsqlDataSource dataSource, JsonSerializerOptions j
         });
     }
 
-    public async Task<IEnumerable<(RecordExpr, float)>> FindOrderByCosineDistanceAsync(float[] queryVector, int skip, int take,
+    public async Task<IEnumerable<(RecordExpr, float)>> FindOrderByCosineDistanceAsync(double[] queryVector, int skip, int take,
         CancellationToken cancellationToken = default)
     {
         await using var conn = dataSource.CreateConnection();
